refactor(frontend): :recycle: move items to recipe folder and break apart types
