+++ frontend/api/class-interfaces/recipes/index.ts
@@ -1,0 +1,1 @@
+export { RecipeAPI } from "./recipe";
